+++ backend/apis/record.py
@@ -30,9 +30,34 @@ class APIRecordHandler(base.BaseHandler):
     async def _query_post(self):
         print('query = ', self.args)
         res = await self.db.getObject('records', cur_user=self.get_current_user_object(), **self.args)
+        cur_user = await self.get_current_user_object()
         for js in res:
             timepoint = int(js['submit_time'].timestamp())
             js['submit_time'] = timepoint
+
+            # authority check
+            if cur_user['role'] == 1:
+                if not js['user_id']==cur_user['id']:
+                    res.remove(js)
+                elif (js['record_type'] == 2 or js['record_type'] == 4) and js['score_openness'] == 0:
+                    self.property_filter(js, None, ['score', 'result', 'consume_time', 'consume_memory', 'status'])
+            elif cur_user['role'] == 2:
+                if js['record_type'] == 0 and js['user_id'] != cur_user['id']:
+                    res.remove(js)
+                elif js['record_type'] == 1 or js['record_type'] == 2 or js['record_type'] == 4:
+                    course = (await self.db.getObject('courses', id=self.args['course_id']))[0]
+                    if not course['id'] in cur_user['ta_courses']:
+                        res.remove(js)
+                elif js['record_type'] == 3:
+                    problem = (await self.db.getObject('problems', id=self.args['problem_id']))[0]
+                    if not problem['user_id'] == cur_user['id']:
+                        res.remove(js)
+            elif cur_user['role'] == 3:
+                pass
+            else:
+                res.remove(js)
+            # ---------------------------------------------------------------------
+
         return res
 
     # @tornado.web.authenticated
@@ -52,6 +77,12 @@ class APIRecordHandler(base.BaseHandler):
     async def _delete_post(self):
         # for condition in self.args:
         res_dict = {}
+        # authority check
+        role = (await self.get_current_user_object())['role']
+        if role < 3:
+            self.set_res_dict(res_dict, code=1, msg='you are not allowed')
+            return res_dict
+
         await self.db.deleteObject('records', **self.args)
         self.set_res_dict(res_dict, code=0, msg='record deleted')
         return res_dict
@@ -59,6 +90,12 @@ class APIRecordHandler(base.BaseHandler):
     # @tornado.web.authenticated
     async def _create_post(self):
         res_dict = {}
+        # authority check
+        role = (await self.get_current_user_object())['role']
+        if role < 3:
+            self.set_res_dict(res_dict, code=1, msg='you are not allowed to use this')
+            return res_dict
+
         await self.db.createObject('records', **self.args)
         self.set_res_dict(res_dict, code=0, msg='record created')
         return res_dict
@@ -77,7 +114,7 @@ class APIRecordHandler(base.BaseHandler):
                        'Compile Error': 8,
                        'unknown': 9,
                        }
-        print('returnresult: ', match_record)               
+        print('returnresult: ', match_record)
         judge_result = self.args['res']
         if match_record['src_language'] == 1 or match_record['src_language'] == 2 or match_record['src_language'] == 4:
             match_record['consume_time'] = judge_result['time']
